Extract the daemon image cache to its own package
ImageCache is now independent of `Daemon` and is located in
`image/cache` package.
